requirements.txt
flask
requests
beautifulsoup4
selenium
google-generativeai